Centralize constants across test cases (#17)
* Centralize default timeout constant

I move the GLOBALTIMEOUT into the constant files and rename it to
DEFAULT_CONSTANT_IN_MS so we can maintain it in one place. I leave
the non-default timeout variable as it is, especially in Freight BI
test cases.

* Centralize dashboard default timeout constant

I move the Dashboard's GLOBALTIMEOUT into the constant files and
rename it to DASHBOARD_CONSTANT_IN_MS so we can maintain it in one
place. This only affects the Freight BI related test cases.

I also rename the advanceFilters filename to fix the import.

* Rename QA automation ID constant

I change the serialnum or QA automation identifier to QA_AUTOMATION_ID
to follow the constant naming and made it clearer.

* Run npm run format

* Cleanup some timeout constant